requirements.txt
firecrawl-py==0.0.12
python-dotenv==1.0.0
flask==2.3.3
requests==2.31.0
sendgrid==6.10.0
